+++ projectroles/tests/test_templatetags.py
@@ -455,40 +455,39 @@ class TestProjectrolesTemplateTags(TestTemplateTagsBase):
         """Test allow_project_creation() in target mode"""
         self.assertEqual(tags.allow_project_creation(), False)
 
-    def test_is_app_link_visible(self):
-        """Test is_app_link_visible()"""
+    def test_is_app_visible(self):
+        """Test is_app_visible()"""
         app_plugin = get_app_plugin('filesfolders')
         self.assertEqual(
-            tags.is_app_link_visible(app_plugin, self.project, self.user), True
+            tags.is_app_visible(app_plugin, self.project, self.user), True
         )
 
-    def test_is_app_link_visible_category(self):
-        """Test is_app_link_visible() with a category"""
+    def test_is_app_visible_category(self):
+        """Test is_app_visible() with a category"""
         app_plugin = get_app_plugin('filesfolders')
         self.assertEqual(
-            tags.is_app_link_visible(app_plugin, self.category, self.user),
-            False,
+            tags.is_app_visible(app_plugin, self.category, self.user), False,
         )
 
-    def test_is_app_link_visible_category_enabled(self):
-        """Test is_app_link_visible() with category_enable=True"""
+    def test_is_app_visible_category_enabled(self):
+        """Test is_app_visible() with category_enable=True"""
         app_plugin = get_app_plugin('timeline')
         self.assertEqual(
-            tags.is_app_link_visible(app_plugin, self.category, self.user), True
+            tags.is_app_visible(app_plugin, self.category, self.user), True
         )
 
     @override_settings(PROJECTROLES_HIDE_APP_LINKS=['filesfolders'])
-    def test_is_app_link_visible_hide(self):
-        """Test is_app_link_visible() with a hidden app and normal/superuser"""
+    def test_is_app_visible_hide(self):
+        """Test is_app_visible() with a hidden app and normal/superuser"""
         app_plugin = get_app_plugin('filesfolders')
         superuser = self.make_user('superuser')
         superuser.is_superuser = True
         superuser.save()
         self.assertEqual(
-            tags.is_app_link_visible(app_plugin, self.project, self.user), False
+            tags.is_app_visible(app_plugin, self.project, self.user), False
         )
         self.assertEqual(
-            tags.is_app_link_visible(app_plugin, self.project, superuser), True
+            tags.is_app_visible(app_plugin, self.project, superuser), True
         )
 
     def test_get_project_list(self):